+++ tybotRoute/engine/mock/MockBotsDataSource.js
@@ -38,7 +38,7 @@ class MockBotsDataSource {
    * @returns an Array of matches
    */
   async getByExactMatch(botId, text) {
-    const intent_display_name = this.data.bots[botId].questions_intent[text];
+    const intent_display_name = this.data.bots[botId].questions_intent?.[text];
     if (intent_display_name) {
       return [this.data.bots[botId].intents[intent_display_name]];
     }