+++ tybotRoute/test/conversation-ai_prompt_bot.js
@@ -108,6 +108,113 @@ const bot = {
 			}]
 		},
 		{
+			"webhook_enabled": false,
+			"enabled": true,
+			"language": "en",
+			"intent_display_name": "ai_prompt_mcp",
+			"intent_id": "00f93b97-89ee-466d-a09c-e47a18943057",
+			"form": {},
+			"question": "",
+			"actions": [{
+				"_tdActionType": "ai_prompt",
+				"_tdActionTitle": "ai action",
+				"assignReplyTo": "ai_reply",
+				"question": "this is the question",
+				"servers": [ { name: "email", transport: "streamable_http", url: "example_url1.com/mcp" }, { name: "calendar", transport: "streamable_http", url: "example_url2.com/mcp" } ],
+				"llm": "myllm",
+				"model": "llmmodel",
+				"max_tokens": 512,
+				"temperature": 0.7,
+				"trueIntent": "#SUCCESS",
+				"falseIntent": "#FAILURE",
+			}]
+		},
+		{
+			"webhook_enabled": false,
+			"enabled": true,
+			"language": "en",
+			"intent_display_name": "ai_prompt_mcp",
+			"intent_id": "00f93b97-89ee-466d-a09c-e47a18943057",
+			"form": {},
+			"question": "",
+			"actions": [{
+				"_tdActionType": "ai_prompt",
+				"_tdActionTitle": "ai action",
+				"assignReplyTo": "ai_reply",
+				"question": "this is the question",
+				"servers": [ { name: "email", transport: "streamable_http", url: "example_url1.com/mcp" }, { name: "calendar", transport: "streamable_http", url: "example_url2.com/mcp" } ],
+				"llm": "myllm",
+				"model": "llmmodel",
+				"max_tokens": 512,
+				"temperature": 0.7,
+				"trueIntent": "#SUCCESS",
+				"falseIntent": "#FAILURE",
+			}]
+		},
+		{
+			"webhook_enabled": false,
+			"enabled": true,
+			"language": "en",
+			"intent_display_name": "ai_prompt_mcp_tools_list",
+			"intent_id": "00f93b97-89ee-466d-a09c-e47a18943057",
+			"form": {},
+			"question": "",
+			"actions": [{
+				"_tdActionType": "ai_prompt",
+				"_tdActionTitle": "ai action",
+				"assignReplyTo": "ai_reply",
+				"question": "this is the question",
+				"servers": [ 
+					{ 
+						name: "email", 
+						transport: "streamable_http", 
+						url: "example_url1.com/mcp", 
+						enabled_toold: ['email_send', 'email_read'] 
+					}, 
+					{ 
+						name: "calendar",
+						transport: "streamable_http", 
+						url: "example_url2.com/mcp", 
+						enabled_toold: ['calendar_read', 'calendar_write'] 
+					},
+					{ 
+						name: "custom",
+						transport: "streamable_http", 
+						url: "example_customurl1.com/mcp", 
+						enabled_toold: ['tool1', 'tool2'] 
+					}
+				],
+				"llm": "myllm",
+				"model": "llmmodel",
+				"max_tokens": 512,
+				"temperature": 0.7,
+				"trueIntent": "#SUCCESS",
+				"falseIntent": "#FAILURE",
+			}]
+		},
+		{
+			"webhook_enabled": false,
+			"enabled": true,
+			"language": "en",
+			"intent_display_name": "ai_prompt_internal_mcp",
+			"intent_id": "00f93b97-89ee-466d-a09c-e47a18943057",
+			"form": {},
+			"question": "",
+			"actions": [{
+				"_tdActionType": "ai_prompt",
+				"_tdActionTitle": "ai action",
+				"assignReplyTo": "ai_reply",
+				"question": "this is the question",
+				"attach": "https://repo.com/example_image.png",
+				"llm": "myllm",
+				"model": "llmmodel",
+				"max_tokens": 512,
+				"temperature": 0.7,
+				"trueIntent": "#SUCCESS",
+				"falseIntent": "#FAILURE",
+			}]
+		},
+		{
 			// TRUE INTENT
 			"webhook_enabled": false,
 			"enabled": true,