+++ tybotRoute/test/conversation-form-test.js
@@ -205,7 +205,7 @@ describe('Conversation1 - Form filling', async () => {
           else {
             assert(params);
             assert(params["last_message_id"] === message_id);
-          assert(params["project_id"] === PROJECT_ID);
+            assert(params["project_id"] === PROJECT_ID);
             assert(params["your_fullname"] === reply_text);
             assert(params["_tdTypeOf:your_fullname"]);
             listener.close(() => {