+++ tybotRoute/tiledeskChatbotPlugs/directives/DirDeleteVariable.js
@@ -1,5 +1,4 @@
 const { param } = require('express/lib/request');
-const ms = require('minimist-string');
 const { TiledeskChatbot } = require('../../engine/TiledeskChatbot');
 const { Filler } = require('../Filler');
 const winston = require('../../utils/winston');
@@ -14,7 +13,7 @@ class DirDeleteVariable {
     this.context = context;
     this.requestId = this.context.requestId;
 
-    this.logger = new Logger({ request_id: this.requestId, dev: this.context.supportRequest?.draft, intent_id: this.context.reply?.attributes?.intent_info?.intent_id });
+    this.logger = new Logger({ request_id: this.requestId, dev: this.context.supportRequest?.draft, intent_id: this.context.reply?.intent_id || this.context.reply?.attributes?.intent_info?.intent_id });
   }
 
   async execute(directive, callback) {