+++ tybotRoute/tiledeskChatbotPlugs/directives/DirDepartment.js
@@ -13,7 +13,7 @@ class DirDepartment {
     this.API_ENDPOINT = context.API_ENDPOINT;
     this.log = context.log;
     
-    this.logger = new Logger({ request_id: this.requestId, dev: this.context.supportRequest?.draft, intent_id: this.context.reply?.attributes?.intent_info?.intent_id });
+    this.logger = new Logger({ request_id: this.requestId, dev: this.context.supportRequest?.draft, intent_id: this.context.reply?.intent_id || this.context.reply?.attributes?.intent_info?.intent_id });
     this.tdClient = new TiledeskClient({ projectId: this.context.projectId, token: this.context.token, APIURL: this.API_ENDPOINT, APIKEY: "___" });
   }