+++ tybotRoute/tiledeskChatbotPlugs/directives/DirIntent.js
@@ -1,7 +1,6 @@
 let axios = require('axios');
 let https = require("https");
 const { v4: uuidv4 } = require('uuid');
-const ms = require('minimist-string');
 const winston = require('../../utils/winston');
 const tilebotService = require('../../services/TilebotService');
 
@@ -71,6 +70,10 @@ class DirIntent {
       "token": this.token
     }
     winston.debug("DirIntent move to intent message: ", intent_command_request);
+
+    // tilebotService.executeBlock(intent_command_request, botId, () => {
+    //   callback(true);
+    // });
 
     tilebotService.sendMessageToBot(intent_command_request, botId, () => {
       callback(true);