+++ tybotRoute/tiledeskChatbotPlugs/directives/DirRandomReply.js
@@ -18,7 +18,7 @@ class DirRandomReply {
     this.tdcache = context.tdcache;
     this.API_ENDPOINT = context.API_ENDPOINT;
     
-    this.logger = new Logger({ request_id: this.requestId, dev: this.context.supportRequest?.draft, intent_id: this.context.reply?.attributes?.intent_info?.intent_id });
+    this.logger = new Logger({ request_id: this.requestId, dev: this.context.supportRequest?.draft, intent_id: this.context.reply?.intent_id || this.context.reply?.attributes?.intent_info?.intent_id });
     this.tdClient = new TiledeskClient({ projectId: this.context.projectId, token: this.context.token, APIURL: this.API_ENDPOINT, APIKEY: "___" });
   }